+++ oap-server/pom.xml
@@ -78,7 +78,7 @@
         <simpleclient.version>0.6.0</simpleclient.version>
 
         <maven-docker-plugin.version>0.28.0</maven-docker-plugin.version>
-        <apollo.version>1.3.0</apollo.version>
+        <apollo.version>1.4.0</apollo.version>
     </properties>
 
     <dependencies>
