+++ oap-server/server-configuration/configuration-apollo/src/test/java/org/apache/skywalking/oap/server/configuration/apollo/ApolloConfigurationTestProvider.java
@@ -16,7 +16,7 @@
  *
  */
 
-package org.apache.skywalking.oap.server.configuration.nacos;
+package org.apache.skywalking.oap.server.configuration.apollo;
 
 import org.apache.skywalking.oap.server.configuration.api.ConfigChangeWatcher;
 import org.apache.skywalking.oap.server.configuration.api.ConfigurationModule;
