+++ oap-server/server-configuration/configuration-apollo/src/test/resources/META-INF/services/org.apache.skywalking.oap.server.library.module.ModuleProvider
@@ -16,4 +16,4 @@
 #
 #
 
-org.apache.skywalking.oap.server.configuration.nacos.ApolloConfigurationTestProvider
+org.apache.skywalking.oap.server.configuration.apollo.ApolloConfigurationTestProvider
